+++ crates/swc_ecma_transforms_module/tests/common_js.rs
@@ -4920,7 +4920,7 @@ test!(
         strict: true,
         strict_mode: true,
         lazy: Lazy::Bool(false),
-        ignore_dynamic: false
+        ..Default::default()
     }),
     issue_1480_1,
     "
@@ -4941,7 +4941,7 @@ test!(
         strict: true,
         strict_mode: true,
         lazy: Lazy::Bool(false),
-        ignore_dynamic: false
+        ..Default::default()
     }),
     issue_1480_2,
     "